requirement.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
tqdm>=4.60.0
opencv-python>=4.5.0
Pillow>=9.0.0
pyyaml>=6.0.3
termcolor >=3.2.0
kornia>=0.7.4


